requirements.txt
fastapi
uvicorn
streamlit
pandas
sqlite3
openpyxl
python-docx
PyPDF2
qdrant-client
google-generativeai
tqdm
python-dotenv